% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{tdClickHandler}
\alias{tdClickHandler}
\title{Handler for a click on a table row}
\usage{
tdClickHandler(id = NULL, fun, ..., eventId = if (stop.propagation)
  "tdClickEvent" else "tdClickEventWithPropagation",
  css.selector = paste0("#", id, " td.data-frame-td"), app = getApp(),
  no.authentication.required = FALSE, stop.propagation = TRUE,
  auto.select = FALSE, remove.sel.row.selector = NULL)
}
\arguments{
\item{id}{id of the HTML img object}

\item{fun}{the handler fun that will be called when the image is clicked}

\item{...}{additional arguments passed to the handler fun}
}
\description{
Handler for a click on a table row
}

