% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdockerstudio.R
\name{viewRDockerStudio}
\alias{viewRDockerStudio}
\title{View RDockerStudio in the RStudio pane}
\usage{
viewRDockerStudio()
}
\description{
Makes sense when being logged-in to an RStudio server instance
that is running on a webserver that has docker containers.
Your user must be able to run docker commands
}

